requirements.txt
ipython
fastapi
uvicorn
elasticsearch==7.13.4
requests
sqlalchemy
pydantic
requests-aws4auth
logging
python-dotenv
boto3
awscli
pytest
pytest-asyncio
httpx
pyyaml
pytest-mock
botocore
psycopg2-binary
pandas
numpy<2.0
